# More tweaks to reduce call stack size