--- a/Jint/Runtime/Interpreter/Statements/JintExpressionStatement.cs
+++ b/Jint/Runtime/Interpreter/Statements/JintExpressionStatement.cs
@@ -1,32 +1,23 @@
 using Esprima.Ast;
 using Jint.Runtime.Interpreter.Expressions;
-using Jint.Runtime.References;
 
-namespace Jint.Runtime.Interpreter.Statements
+namespace Jint.Runtime.Interpreter.Statements;
+
+internal sealed class JintExpressionStatement : JintStatement<ExpressionStatement>
 {
-    internal sealed class JintExpressionStatement : JintStatement<ExpressionStatement>
+    private JintExpression _expression = null!;
+
+    public JintExpressionStatement(ExpressionStatement statement) : base(statement)
     {
-        private JintExpression _expression = null!;
+    }
 
-        public JintExpressionStatement(ExpressionStatement statement) : base(statement)
-        {
-        }
+    protected override void Initialize(EvaluationContext context)
+    {
+        _expression = JintExpression.Build(context.Engine, _statement.Expression);
+    }
 
-        protected override void Initialize(EvaluationContext context)
-        {
-            _expression = JintExpression.Build(context.Engine, _statement.Expression);
-        }
-
-        protected override Completion ExecuteInternal(EvaluationContext context)
-        {
-            var result = _expression.Evaluate(context);
-
-            if (result.Type != ExpressionCompletionType.Reference)
-            {
-                return new Completion(result);
-            }
-
-            return new Completion(CompletionType.Normal, context.Engine.GetValue((Reference) result.Value, true), null, _statement);
-        }
+    protected override Completion ExecuteInternal(EvaluationContext context)
+    {
+        return _expression.GetValue(context);
     }
 }