--- a/Jint/Runtime/Interpreter/Statements/JintStatement.cs
+++ b/Jint/Runtime/Interpreter/Statements/JintStatement.cs
@@ -33,7 +33,7 @@ namespace Jint.Runtime.Interpreter.Statements
             if (_statement.Type != Nodes.BlockStatement)
             {
                 context.LastSyntaxElement = _statement;
-                context.Engine.RunBeforeExecuteStatementChecks(_statement);
+                context.RunBeforeExecuteStatementChecks(_statement);
             }
 
             if (!_initialized)