@lahma 2022-08-31
> What's up with the strin(...)input [21] benchmark?

I'll try to run the numbers at some point.  This was a bit unexpected..