@lahma 2022-09-01
> What's up with the strin(...)input [21] benchmark?

OK it's a change in behavior: 

https://github.com/sebastienros/jint/pull/1277/files#diff-e715e3b319920d94963426777af53d921e568d5a187aa805a835d1ef4e6063dfL59

The function result is no longer eagerly cloned, in this case string builder (`ConcatenatedString`) isn't being materialized. I need to think if there is a case where it could be a problem, on assignment we usually want to ensure that it's properly materialized,.