@sebastienros 2022-09-01
Another point of view is that since it had such a big impact there, maybe there are other places with the same code that could be optimized the same way, as long as it's also valid.