@lahma 2022-09-01
So far I haven't found a reason this wouldn't generally be safe for string in this case. `ConcatenatedString` is only one currently using the `Clone`, I was afraid of `ArgumentsInstance`, which is pooled and can cause trouble (thanks test suite!). One thing to keep an eye on for sure.